Python Script to parse dictionary files / sorting script
Simply export the file to text with quotes as text qualifiers and commas
as the field delimiter. This will parse and print the files in the
needed format to create a dictionary for the giovanni dictionary
creator.
I had trouble getting access to export the files sorted so I wrote the
sort.py to do that quickly.